Rename inst_stream to program_bytes
It's not really an instruction stream, where the instructions must be
processed one after the other. Rather, it's a random-access chunk of
memory containing the program machine code. It doesn't need to be
iterated sequentially as a strea - it can be random access, as in the
case of jumps and branches. Make this change in variable name to
emphasize this point.