Create a unified memory for both code and data
If executing, resize the initial program bytes vector to 1 MB, and
then move it into the CPU state to be used as memory.
Add a condition that checks to see if we are executing past the
original program, and if we are, stop.
Print out the original program size and the current program bytes size.