Move all built binaries into bin86/ and commit to repo
Committing all the built binaries to the repo will enable running
regressions on non-Linux builds. Non-x86-linux machines are unable
to compile the nasm from the 8086 toolchain.
Newer macs are Arm, and cc doesn't compile down to that target.
Windows didn't seem to compile for me, but I don't remember why not.